Rename requests.ts to clients.ts, update imports
